Implemented some code

+++ Infrastructure/Services/LibraryCardService.cs
@@ -3,6 +3,8 @@ using Application.DTOs.LibraryCards;
 using Application.Services.Interfaces;
 using AutoMapper;
 using Domain.Entities;
+using Domain.Exceptions;
+using Microsoft.Extensions.Logging;
 
 
 namespace Infrastructure.Services
@@ -11,15 +13,20 @@ namespace Infrastructure.Services
     {
         private readonly ILibraryCardRepository _libraryCardRepository;
         private readonly IMapper _mapper;
+        private readonly ILogger<LibraryCardService> _logger;
 
-        public LibraryCardService(ILibraryCardRepository libraryCardRepository,IMapper mapper)
+        public LibraryCardService(ILibraryCardRepository libraryCardRepository, IMapper mapper, ILogger<LibraryCardService> logger)
         {
             _libraryCardRepository = libraryCardRepository;
             _mapper = mapper;
+            _logger = logger;
         }
 
         public async Task<GetLibraryCardDTO> CreateAsync(Guid userId)
         {
+            if (userId == Guid.Empty)
+                throw new ArgumentNullException("userId");
+
             var libraryCardToCreate = new LibraryCard
             {
                 UserId = userId,
@@ -29,27 +36,31 @@ namespace Infrastructure.Services
 
             if (await _libraryCardRepository.IsExistsAsync(userId))
             {
-                throw new Exception("Library card already exists for this user.");
+                throw new LibraryCardExistsException($"User with id {userId} has library card");
             }
 
             var createdLibraryCard = await _libraryCardRepository.CreateAsync(libraryCardToCreate);
 
             if (createdLibraryCard == null)
             {
-                throw new Exception("Failed to create library card.");
+                throw new InvalidOperationException("Failed to create library card.");
             }
 
+            _logger.LogInformation("Library card successfully created");
             return _mapper.Map<GetLibraryCardDTO>(createdLibraryCard);
         }
 
 
         public async Task<bool> DeleteAsync(Guid userId)
         {
+            if (userId == Guid.Empty)
+                throw new ArgumentNullException("userId");
+
             var libraryCardToDelete = await _libraryCardRepository.GetByUserIdAsync(userId);
 
             if (libraryCardToDelete == null)
             {
-                throw new Exception("Library card not found.");
+                throw new LibraryCardNotFoundException($"Library card with userId {userId} not found.");
             }
 
             var result = await _libraryCardRepository.DeleteAsync(libraryCardToDelete.Id);
@@ -58,6 +69,8 @@ namespace Infrastructure.Services
             {
                 throw new Exception("Failed to delete library card.");
             }
+
+            _logger.LogInformation("Library card successfully deleted");
             return true;
         }
 
@@ -67,7 +80,8 @@ namespace Infrastructure.Services
 
             if (libraryCards == null || !libraryCards.Any())
             {
-                throw new Exception("No library cards found.");
+                _logger.LogWarning("No library cards found.");
+                return Enumerable.Empty<GetLibraryCardDTO>();
             }
 
             return _mapper.Map<IEnumerable<GetLibraryCardDTO>>(libraryCards);
@@ -75,11 +89,14 @@ namespace Infrastructure.Services
 
         public async Task<DateTime?> GetExpirationDateAsync(Guid userId)
         {
+            if (userId == Guid.Empty)
+                throw new ArgumentNullException(nameof(userId));
+
             var libraryCard = await _libraryCardRepository.GetByUserIdAsync(userId);
 
             if (libraryCard == null)
             {
-                throw new Exception("Library card not found.");
+                throw new LibraryCardNotFoundException($"Library card with userId {userId} not found.");
             }
 
             return libraryCard.ValidTo;
@@ -87,22 +104,28 @@ namespace Infrastructure.Services
 
         public async Task<bool> IsActiveAsync(Guid userId)
         {
+            if(userId == Guid.Empty)
+                throw new ArgumentNullException(nameof(userId));
+
             var libraryCard = await _libraryCardRepository.GetByUserIdAsync(userId);
 
             if (libraryCard == null)
             {
-                throw new Exception("Library card not found.");
+                throw new LibraryCardNotFoundException($"Library card with userId {userId} not found.");
             }
 
             return libraryCard.IsValid && libraryCard.ValidTo > DateTime.UtcNow;
         }
         public async Task<GetLibraryCardDTO> UpdateAsync(UpdateLibraryCardDTO updateLibraryCardDTO)
         {
+            if (updateLibraryCardDTO == null)
+                throw new ArgumentNullException(nameof(updateLibraryCardDTO));
+
             var libraryCard = await _libraryCardRepository.GetByUserIdAsync(updateLibraryCardDTO.UserId);
 
             if (libraryCard == null)
             {
-                throw new Exception("Library card not found.");
+                throw new LibraryCardNotFoundException($"Library card with userId {updateLibraryCardDTO.UserId} not found.");
             }
 
             _mapper.Map(updateLibraryCardDTO, libraryCard);
@@ -111,9 +134,10 @@ namespace Infrastructure.Services
 
             if (updatedLibraryCard == null)
             {
-                throw new Exception("Failed to update library card.");
+                throw new InvalidOperationException("Failed to update library card.");
             }
 
+            _logger.LogInformation("Successfully updated library card");
             return _mapper.Map<GetLibraryCardDTO>(updatedLibraryCard);
         }
     }
